bug: filter by timestamp (#548)
This includes a filter by cell timestamp. This also includes a lot of
inline comments pointing out potential hazards and pitfalls that I
experienced working on this.

NOTE: There is a (hopefully remote) possibility that a record could be
returned with a specific cell missing if that cell's timestamp has
expired. Timestamps are per cell, not per row.

Closes: SYNC-4060

+++ autopush-common/src/db/mod.rs
@@ -41,7 +41,7 @@ use crate::util::timing::{ms_since_epoch, sec_since_epoch};
 use models::{NotificationHeaders, RangeKey};
 
 const MAX_EXPIRY: u64 = 2_592_000;
-pub const USER_RECORD_VERSION: u8 = 1;
+pub const USER_RECORD_VERSION: u64 = 1;
 /// The maximum TTL for channels, 30 days
 pub const MAX_CHANNEL_TTL: u64 = 30 * 24 * 60 * 60;
 
@@ -166,7 +166,7 @@ pub struct User {
     pub node_id: Option<String>,
     /// Record version
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub record_version: Option<u8>,
+    pub record_version: Option<u64>,
     /// LEGACY: Current month table in the database the user is on
     #[serde(skip_serializing_if = "Option::is_none")]
     pub current_month: Option<String>,
